[consensus] Finish refactor of consensus
