+++ consensus/modified.og
@@ -1,0 +1,12 @@
+accept message? ANNOUNCE false Normal false false true
+accept message? ANNOUNCE false Normal false false true
+accept message? ANNOUNCE false Normal false false true
+accept message? ANNOUNCE false Normal false false true
+accept message? ANNOUNCE false Normal false false true
+accept message? ANNOUNCE false Normal false false true
+accept message? ANNOUNCE false Normal false false true
+accept message? ANNOUNCE false Normal false false true
+accept message? ANNOUNCE false Normal false false true
+accept message? ANNOUNCE false Normal false false true
+accept message? ANNOUNCE false Normal false false true
+accept message? ANNOUNCE false Normal false false true
